python/packages/ai/pyproject.toml
[tool.poetry]
name = "teams-ai"
version = "1.0.1"
description = "SDK focused on building AI based applications for Microsoft Teams."
authors = ["Microsoft <teams@microsoft.com>"]
readme = "README.md"
repository = "https://github.com/microsoft/teams-ai"
documentation = "https://learn.microsoft.com/en-us/microsoftteams/platform/bots/how-to/teams%20conversational%20ai/teams-conversation-ai-overview"
keywords = ["microsoft", "teams", "ai", "bot"]
packages = [
    { include = "teams" },
    { include = "tests" }
]

[tool.poetry.dependencies]
botbuilder-core = "^4.14.8"
botbuilder-integration-aiohttp = "^4.14.8"
botframework-connector = "^4.14.8"
python = ">=3.8,<4.0"
tiktoken = "^0.4.0"
aiohttp = "^3.8.5"
jsonschema = "^4.21.1"
types-pyyaml = "^6.0.12.12"
pyyaml = "^6.0.1"
dataclasses-json = "^0.6.4"
openai = "^1.11.1"
azure-ai-contentsafety = "^1.0.0"
msal = "^1.28.0"
botbuilder-dialogs = "^4.14.8"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pylint = "^2.17.4"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"
black = ">=23.7,<25.0"
isort = "^5.12.0"
mypy = "^1.5.0"
httpx = "^0.26.0"

[tool.poetry.scripts]
lint = "scripts:lint"
fmt = "scripts:fmt"
test = "scripts:test"
clean = "scripts:clean"
ci = "scripts:ci"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
addopts = "-rx --cov-report html:coverage/html --cov-report lcov:coverage/lcov.info --cov=teams"
log_cli = true
log_cli_level = "INFO"

[tool.mypy]
python_version = "3.8"
ignore_missing_imports = true
show_error_codes = true
no_implicit_optional = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "tests.*"
disable_error_code = ["var-annotated", "method-assign"]
